reorganised test files
